requirements.txt
asgiref==3.4.1
Django==3.2.8
django-admin-sortable2==1.0.2
django-resized==0.3.11
django-ckeditor==6.1.0
django-colorfield==0.6.0
django-js-asset==1.2.2
django-modeltranslation==0.18.5
django-unused-media==0.2.2
django-imagekit==4.0.2
gunicorn==21.2.0
Pillow==8.4.0
python-monkey-business==1.0.0
pilkit==2.0
pytils==0.3
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.0.1